feat(ssm): deterministic fixed-point SSM working memory with salience/novelty and attention feedback
- Provide a compact, deterministic integer fixed-point State Space Model (SSM) working memory as a SelectiveScan core to close the perception→memory→attention feedback loop.
- Expose salience, novelty and an attention_gain suggestion so downstream modules (FeatureSpiker, Attention, Workspace, Archive, CDE) can deterministically react to working memory state.
- Keep the implementation lightweight and dependency-free (no ML libs / protobufs) and within strict numeric bounds for safety and reproducibility.

- Added a new deterministic SSM core at `core/crates/ucf-ssm/src/lib.rs` with public types `SsmParams`, `SsmState`, `SsmInputs`, `SsmOutputs` and `SsmCore` implementing `fn tick(&mut self, inp: &SsmInputs) -> SsmOutputs` using fixed-point arithmetic and a sequential Selective Scan update.
- Implemented compact deterministic control vector construction (hashed signed components), per-dimension update with leak/attention modulation and clamping, and deterministic commits/digests (`commit_state`, `digest_state`, `commit_outputs`, `commit_core`).
- Computed `salience`, `novelty` (digest xor popcount → 0..10000) and `attention_gain` (weighted salience/novelty with risk/drift inhibition) in the SSM output.
- Wired SSM into the router and workspace: `core/crates/ucf-router/src/lib.rs` now builds `SsmInputs` from percept + NCDE + spikes and calls `SsmCore::tick`, stores full snapshot into `Workspace` via `set_ssm_snapshot`, and the attention combination rule is implemented in `apply_ssm_attention_bias` as `max(policy_gain, ssm.attention_gain)` to deterministically bias attention.
- Updated workspace and archive plumbing to carry `ssm_commit`, `ssm_state_commit`, `ssm_salience`, `ssm_novelty`, and `ssm_attention_gain` (fields and encode/commit functions in `core/crates/ucf-workspace/src/lib.rs`), plus archive record payload changes and SSM output record formatting.
- Adjusted structural tuning to target `ssm.k_novelty` (renamed param) and propagated changes where structural params are applied (`core/crates/ucf-structural-store` and `core/crates/ucf-rsa`).
- Integration touches: replaced prior `ucf-ssm` legacy API usages in router and tests to the new `SsmInputs`/`SsmOutputs` shapes and removed an unnecessary dev-dependency from `core/crates/ucf-ssm/Cargo.toml`.
- Added/updated unit tests: deterministic output commits, leak decay, novelty sensitivity to percept changes, attention_gain monotonicity, and a `feature_spiker` test to assert SSM metrics drive spike production (`domains/features/crates/ucf-feature-spiker`).

- Ran the full test suite with `cargo test -q`, which completed with all tests passing after the changes.
- Exercised SSM unit tests that verify determinism (`commit`/`ssm_state_commit` stable), decay behavior for leak, novelty increase on percept change, and attention_gain scaling bounds.
- Ran integration tests that validate router→workspace→attention wiring and `FeatureSpiker` behavior using the new SSM metrics, all of which passed.

fix(ssm): resolve clippy warning for scale constant

+++ core/crates/ucf-ssm/src/lib.rs
@@ -227,7 +227,7 @@ fn build_control_vector(inp: &SsmInputs, dim: usize) -> Vec<i64> {
     let spike_mix = i64::from(feature)
         .saturating_add(i64::from(novelty).saturating_mul(2))
         .saturating_add(i64::from(threat).saturating_mul(3))
-        .min(SCALE as i64);
+        .min(SCALE);
     let energy_mix = i64::from(inp.ncde_energy).saturating_add(i64::from(inp.percept_energy) / 2);
     let attn_mix = i64::from(inp.prev_attention_gain) / 3;
     let bias = spike_mix
